requirements.txt
channels==0.17.2
Django==1.10.2
django-cors-middleware==1.3.1
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
